HIMAN-257: optimizing preform_hybrid
preform_hybrid examines the properties of the zero levels below 10km.
Initial implementation follow smarttool logic where all levels were
processed separately. This is not efficient when grid size is large.

This commit adds modifier the functionality that it returns all found
values in the given height range. This is triggered with 'findNth' value
-1. Default behavior is not changed and test results should not be
affected.

+++ himan-lib/include/modifier.h
@@ -50,8 +50,8 @@ class modifier
 
 	virtual bool CalculationFinished() const;
 
-	size_t FindNth() const;
-	virtual void FindNth(size_t theNth);
+	int FindNth() const;
+	virtual void FindNth(int theNth);
 
 	virtual void Process(const std::vector<double>& theData, const std::vector<double>& theHeights);
 
@@ -77,7 +77,7 @@ class modifier
 
 	/**
 	 * @brief Initialize lower and upper heights to some default values
-	     */
+	 */
 
 	virtual void InitializeHeights();
 
@@ -92,7 +92,7 @@ class modifier
 	std::vector<double> itsPreviousValue;
 	std::vector<double> itsPreviousHeight;
 
-	size_t itsFindNthValue;
+	int itsFindNthValue;
 
 	mutable std::vector<double> itsResult;  // variable is modified in some Result() const functions
 	size_t itsIndex;
@@ -333,7 +333,7 @@ class modifier_findheight : public modifier
 	}
 	virtual void Init(const std::vector<double>& theData, const std::vector<double>& theHeights) override;
 
-	std::vector<size_t> itsFoundNValues;
+	std::vector<int> itsFoundNValues;
 
 	size_t itsValuesFound;
 };
@@ -357,7 +357,7 @@ class modifier_findheight_gt : public modifier_findheight
 	}
 	virtual void Calculate(double theValue, double theHeight, double thePreviousValue,
 	                       double thePreviousHeight) override;
-	virtual void FindNth(size_t theNth) override;
+	virtual void FindNth(int theNth) override;
 };
 
 /**
@@ -379,7 +379,7 @@ class modifier_findheight_lt : public modifier_findheight
 	}
 	virtual void Calculate(double theValue, double theHeight, double thePreviousValue,
 	                       double thePreviousHeight) override;
-	virtual void FindNth(size_t theNth) override;
+	virtual void FindNth(int theNth) override;
 };
 
 /**
